refactor: modularize root command and extract utility functions
- Move `runUpdatesWithNotifications` function to `actions.RunUpdatesWithNotifications`
- Extract API setup logic to `internal/api.SetupAndStartAPI`
- Move scheduling logic to `internal/scheduling.RunUpgradesOnSchedule`
- Extract logging utilities to `internal/logging.WriteStartupMessage`
- Move configuration struct to `internal/config.RunConfig`
- Relocate utility functions (`FormatDuration`, `FormatTimeUnit`, `FilterEmpty`) to `internal/util`
- Move report types (`SingleContainerReport`, `SortableContainers`) to `pkg/session/report.go`
- Update notifier interface with new methods for filtered entries
- Adjust tests to use new module functions and types
- Remove unused imports and clean up code structure in `cmd/root.go
