+++ cmd/root_test.go
@@ -19,7 +19,11 @@ import (
 
 	dockerContainer "github.com/docker/docker/api/types/container"
 
+	"github.com/nicholas-fedor/watchtower/internal/api"
 	"github.com/nicholas-fedor/watchtower/internal/flags"
+	"github.com/nicholas-fedor/watchtower/internal/logging"
+	"github.com/nicholas-fedor/watchtower/internal/scheduling"
+	"github.com/nicholas-fedor/watchtower/internal/util"
 	"github.com/nicholas-fedor/watchtower/pkg/api/update"
 	containerMock "github.com/nicholas-fedor/watchtower/pkg/container/mocks"
 	"github.com/nicholas-fedor/watchtower/pkg/metrics"
@@ -364,7 +368,7 @@ func TestFormatDuration(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result := formatDuration(tt.duration)
+			result := util.FormatDuration(tt.duration)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
@@ -415,11 +419,7 @@ func TestFormatTimeUnit(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result := formatTimeUnit(struct {
-				value    int64
-				singular string
-				plural   string
-			}{tt.value, tt.singular, tt.plural}, tt.forceInclude)
+			result := util.FormatTimeUnit(tt.value, tt.singular, tt.plural, tt.forceInclude)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
@@ -455,7 +455,7 @@ func TestFilterEmpty(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result := filterEmpty(tt.input)
+			result := util.FilterEmpty(tt.input)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
@@ -545,7 +545,7 @@ func TestGetAPIAddr(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result := getAPIAddr(tt.host, tt.port)
+			result := api.GetAPIAddr(tt.host, tt.port)
 			assert.Equal(t, tt.expected, result)
 
 			// Verify the formatted address is a valid TCP address
@@ -747,7 +747,21 @@ func TestUpdateOnStartTriggersImmediateUpdate(t *testing.T) {
 	filterDesc := testFilterDesc
 
 	// The function should trigger immediate update and then start scheduler
-	err = runUpgradesOnSchedule(ctx, cmd, filter, filterDesc, updateLock, false)
+	err = scheduling.RunUpgradesOnSchedule(
+		ctx,
+		cmd,
+		filter,
+		filterDesc,
+		updateLock,
+		false,
+		"",
+		logging.WriteStartupMessage,
+		runUpdatesWithNotifications,
+		nil,
+		"",
+		nil,
+		"",
+	)
 
 	// Should not return an error (context cancellation is expected)
 	require.NoError(t, err)
@@ -816,7 +830,21 @@ func TestUpdateOnStartIntegratesWithCronScheduling(t *testing.T) {
 	filterDesc := testFilterDesc
 
 	startTime := time.Now()
-	err = runUpgradesOnSchedule(ctx, cmd, filter, filterDesc, updateLock, false)
+	err = scheduling.RunUpgradesOnSchedule(
+		ctx,
+		cmd,
+		filter,
+		filterDesc,
+		updateLock,
+		false,
+		"",
+		logging.WriteStartupMessage,
+		runUpdatesWithNotifications,
+		nil,
+		"",
+		nil,
+		"",
+	)
 
 	// Should not return an error (context cancellation is expected)
 	require.NoError(t, err)
@@ -888,7 +916,21 @@ func TestUpdateOnStartLockingBehavior(t *testing.T) {
 	filter := func(_ types.FilterableContainer) bool { return false }
 	filterDesc := testFilterDesc
 
-	err = runUpgradesOnSchedule(ctx, cmd, filter, filterDesc, updateLock, false)
+	err = scheduling.RunUpgradesOnSchedule(
+		ctx,
+		cmd,
+		filter,
+		filterDesc,
+		updateLock,
+		false,
+		"",
+		logging.WriteStartupMessage,
+		runUpdatesWithNotifications,
+		nil,
+		"",
+		nil,
+		"",
+	)
 
 	// Should not return an error
 	require.NoError(t, err)
@@ -939,7 +981,21 @@ func TestUpdateOnStartSelfUpdateScenario(t *testing.T) {
 	filter := func(_ types.FilterableContainer) bool { return true }
 	filterDesc := testFilterDesc
 
-	err = runUpgradesOnSchedule(ctx, cmd, filter, filterDesc, updateLock, false)
+	err = scheduling.RunUpgradesOnSchedule(
+		ctx,
+		cmd,
+		filter,
+		filterDesc,
+		updateLock,
+		false,
+		"",
+		logging.WriteStartupMessage,
+		runUpdatesWithNotifications,
+		nil,
+		"",
+		nil,
+		"",
+	)
 
 	// Should not return an error
 	require.NoError(t, err)
@@ -1001,7 +1057,21 @@ func TestUpdateOnStartMultiInstanceScenario(t *testing.T) {
 		filter := func(_ types.FilterableContainer) bool { return false }
 		filterDesc := "instance1"
 
-		err := runUpgradesOnSchedule(ctx, cmd1, filter, filterDesc, updateLock, false)
+		err := scheduling.RunUpgradesOnSchedule(
+			ctx,
+			cmd1,
+			filter,
+			filterDesc,
+			updateLock,
+			false,
+			"",
+			logging.WriteStartupMessage,
+			runUpdatesWithNotifications,
+			nil,
+			"",
+			nil,
+			"",
+		)
 		assert.NoError(t, err)
 		atomic.AddInt32(&completed, 1)
 		close(instance1Called)
@@ -1014,7 +1084,21 @@ func TestUpdateOnStartMultiInstanceScenario(t *testing.T) {
 		filter := func(_ types.FilterableContainer) bool { return false }
 		filterDesc := "instance2"
 
-		err := runUpgradesOnSchedule(ctx, cmd2, filter, filterDesc, updateLock, false)
+		err := scheduling.RunUpgradesOnSchedule(
+			ctx,
+			cmd2,
+			filter,
+			filterDesc,
+			updateLock,
+			false,
+			"",
+			logging.WriteStartupMessage,
+			runUpdatesWithNotifications,
+			nil,
+			"",
+			nil,
+			"",
+		)
 		assert.NoError(t, err)
 		atomic.AddInt32(&completed, 1)
 		close(instance2Called)
@@ -1059,7 +1143,7 @@ func TestWaitForRunningUpdate_NoUpdateRunning(t *testing.T) {
 	ctx := context.Background()
 	start := time.Now()
 
-	waitForRunningUpdate(ctx, lock)
+	scheduling.WaitForRunningUpdate(ctx, lock)
 
 	elapsed := time.Since(start)
 
@@ -1077,7 +1161,7 @@ func TestWaitForRunningUpdate_UpdateRunning(t *testing.T) {
 	waitCompleted := make(chan bool, 1)
 
 	go func() {
-		waitForRunningUpdate(ctx, lock)
+		scheduling.WaitForRunningUpdate(ctx, lock)
 
 		waitCompleted <- true
 	}()
@@ -1141,7 +1225,21 @@ func TestRunUpgradesOnSchedule_ShutdownWaitsForRunningUpdate(t *testing.T) {
 		filterDesc := testFilterDesc
 
 		// This should start and wait for context cancellation
-		err := runUpgradesOnSchedule(ctx, cmd, filter, filterDesc, updateLock, false)
+		err := scheduling.RunUpgradesOnSchedule(
+			ctx,
+			cmd,
+			filter,
+			filterDesc,
+			updateLock,
+			false,
+			"",
+			logging.WriteStartupMessage,
+			runUpdatesWithNotifications,
+			nil,
+			"",
+			nil,
+			"",
+		)
 		assert.NoError(t, err)
 
 		shutdownCompleted <- true
