+++ pkg/session/report_test.go
@@ -831,20 +831,20 @@ func Test_sortCategories(t *testing.T) {
 	}
 }
 
-func Test_sortableContainers_Len(t *testing.T) {
+func Test_SortableContainers_Len(t *testing.T) {
 	tests := []struct {
 		name string
-		s    sortableContainers
+		s    SortableContainers
 		want int
 	}{
 		{
 			name: "empty slice",
-			s:    sortableContainers{},
+			s:    SortableContainers{},
 			want: 0,
 		},
 		{
 			name: "two elements",
-			s: sortableContainers{
+			s: SortableContainers{
 				mocks.NewMockContainerReport(t),
 				mocks.NewMockContainerReport(t),
 			},
@@ -854,13 +854,13 @@ func Test_sortableContainers_Len(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := tt.s.Len(); got != tt.want {
-				t.Errorf("sortableContainers.Len() = %v, want %v", got, tt.want)
+				t.Errorf("SortableContainers.Len() = %v, want %v", got, tt.want)
 			}
 		})
 	}
 }
 
-func Test_sortableContainers_Less(t *testing.T) {
+func Test_SortableContainers_Less(t *testing.T) {
 	type args struct {
 		i int
 		j int
@@ -868,13 +868,13 @@ func Test_sortableContainers_Less(t *testing.T) {
 
 	tests := []struct {
 		name string
-		s    sortableContainers
+		s    SortableContainers
 		args args
 		want bool
 	}{
 		{
 			name: "lower ID first",
-			s: sortableContainers{
+			s: SortableContainers{
 				func() types.ContainerReport {
 					mock := mocks.NewMockContainerReport(t)
 					mock.EXPECT().ID().Return(types.ContainerID("cont1"))
@@ -893,7 +893,7 @@ func Test_sortableContainers_Less(t *testing.T) {
 		},
 		{
 			name: "higher ID first",
-			s: sortableContainers{
+			s: SortableContainers{
 				func() types.ContainerReport {
 					mock := mocks.NewMockContainerReport(t)
 					mock.EXPECT().ID().Return(types.ContainerID("cont2"))
@@ -914,13 +914,13 @@ func Test_sortableContainers_Less(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := tt.s.Less(tt.args.i, tt.args.j); got != tt.want {
-				t.Errorf("sortableContainers.Less() = %v, want %v", got, tt.want)
+				t.Errorf("SortableContainers.Less() = %v, want %v", got, tt.want)
 			}
 		})
 	}
 }
 
-func Test_sortableContainers_Swap(t *testing.T) {
+func Test_SortableContainers_Swap(t *testing.T) {
 	type args struct {
 		i int
 		j int
@@ -928,13 +928,13 @@ func Test_sortableContainers_Swap(t *testing.T) {
 
 	tests := []struct {
 		name string
-		s    sortableContainers
+		s    SortableContainers
 		args args
 		want []string // Use IDs directly instead of mocks for want
 	}{
 		{
 			name: "swap first and second",
-			s: sortableContainers{
+			s: SortableContainers{
 				func() types.ContainerReport {
 					mock := mocks.NewMockContainerReport(t)
 					// Expect 1 call in the comparison loop after swap
@@ -958,7 +958,7 @@ func Test_sortableContainers_Swap(t *testing.T) {
 			tt.s.Swap(tt.args.i, tt.args.j)
 
 			if len(tt.s) != len(tt.want) {
-				t.Errorf("sortableContainers.Swap() length = %d, want %d", len(tt.s), len(tt.want))
+				t.Errorf("SortableContainers.Swap() length = %d, want %d", len(tt.s), len(tt.want))
 
 				return
 			}
@@ -969,7 +969,7 @@ func Test_sortableContainers_Swap(t *testing.T) {
 				gotID := tt.s[i].ID()
 				if gotID != types.ContainerID(tt.want[i]) {
 					t.Errorf(
-						"sortableContainers.Swap()[%d].ID() = %v, want %v",
+						"SortableContainers.Swap()[%d].ID() = %v, want %v",
 						i,
 						gotID,
 						tt.want[i],
